Where You At attempt 2
Due to a titanium crash had to start the app over again as it corrupted
the entire project.  Got the main interface created and maps working on
both the iOS simulator and an actual android device and tested on them
going to test the android emulator and the actual iPad.  Not to sure if
the android emulator will work as titanium says that the map module
will not work with the emulator at all.