Where You At
restarted the project as a titanium crash corrupted the entire previous
version of the app.  Got the map module and the main UI created for the
app.  Map is working on both android devices as well as the iOs
simulator will be testing on iPad and android emulator though the
titanium documentation says that the map module will not work on the
android emulator